Added pain accumulation system
Usage is similar to PainChance:
* Actor has either MaxPain [int] or MaxPain "DamageType", [int].
* PainAmount [int] on the damage inflictor determines how much pain to add.
* When the threshold for the given damage type is reached, Actor enters pain state and accumulation is reset.

Note that the accumulation amount is shared across all damage types, only the thresholds differ. If the default MaxPain is 0, the threshold system is ignored in place of the old chance system (if a damage type is provided with MaxPain, it will still check against it, but only for those damage types).